Promote NegotiateSession to first-class tuyaAPI method
Add a NegotiateSession() method. For the protocols before 3.4 which don't
do actual session negotiation, this just sets the encryption key.

For the 3.4 protocol it does the full negotiation on the wire as it did
before, now under the explicit control of the user instead of implicitly
blocking in the first BuildTuyaMessage() call.

This change also removes the encryption_key parameter from BuildTuyaMessage()
and DecodeTuyaMessage() since the key is now stored in a member variable.

The examples now call NegotiateSession(device_key) after ConnectToDevice().